requirements.txt
Flask==2.0.1
psycopg2-binary==2.9.1
python-dotenv==0.18.0
gunicorn==20.1.0
PyYAML==5.4.1
redis==3.5.3
